Crystalline / Organizing Files 0.01  Update
Created a Separate Branch for testing of diseased objects.

Included previous code from previous project, will delete one coral to ensure object is no-longer instantiated along with working on the project and commenting most of the code. 

Will continue to work Corals to produce better effects.

Created folders in addition to make biome specific and coding specialties that are different from peers. 

Also included a different vertice called CrystalVert which will be used to make emissive verticies specific to the Crystal Biome.

Added in two Game Objects / Prefabs that control a Flower and A Crystalline Object. They are not supposed to be organice, but instead inorganic material that mimics organic creatures.

Added in an Addition Scene for testing of coral.